fix: A* floating point on parser
Former-commit-id: 7f674b0105008663f19a32b4297d3b911eb787d2

+++ BestPath/Models/Coordinates.cs
@@ -1,3 +1,3 @@
 namespace BestPath.Models;
 
-public record Coordinate(int x, int y);
+public record Coordinate(float x, float y);
